feature: multiple root paths
This feature allows to pass an array as `path` argument to create an
overlay. The first matched file will be used for size and modified date.

The new option `caseSensitive` allows to specify if case-insensitive duplicates should be listed.

The response functions (html/json/plain) will be called with a different
set of parameters
- `next` was moved on 3rd place to align with default middleware
functions
- `files` array contains an object (name,path) instead of just the name
to allow accessing the files as they can be located in different dirs
- removed `path` as it is contained in files and can be different for
each file

Most of the tests were changed to run with this feature by splitting up
the `fixtures` into two sub-folders. There are some duplicate files by
intent to test that only the first match will be used and no duplicates
are served as index.

fixes https://github.com/expressjs/serve-index/issues/24